Added in Google Play Services check
Error handling is currently commented out because the emulators dont
have play services installed and cant even attempt to install it. I am
working on getting a development phone so I can test the error handling.
The play services have to be current in order to do what we plan to do
with the users account.