Updated layout of MapData (need to rename to something more relevent)
Updated the screen so there is only 1 butt on for stoping and starting a
workout. Since there is no longer a login webview I tried to fill the
empty space the best I could at this stage. It will eventually be filled
with workout options/data. On this page I also updated the toast
messages that are displayed.